Split pointer event handling into a separate file
